fix: remove providers in items and collections (#6)
Providers are not a required field in STAC spec.

Besides, what we currently have in EODAG is the backend which is already
set in federation:backend field.
The added value would be to have providers with roles such as producer,
licensor or processor, but it is not yet possible with EODAG

Signed-off-by: Aubin Lambaré <aubin.lambare@cs-soprasteria.com>