fix: Fixed free-text search (#48)
Even though `stac-fastapi-eodag` defined the `q` parameter as an
optional string in `all_collections`, this seems to be ignored by
fastapi since the `q` is defined by the module `stac-fastapi` on the
free_text extension as a list of strings.

This requires the `q` parameter to be expanded, using AND by default,
which can be parsed by the eodag free-text parser.

---------

Co-authored-by: Guillermo Tesoro Calvo <guillermo.tesoro@c-ssystems.de>